requirements-dev.txt

pytest
pytest-cov
pytest-xprocess
flake8

#sarge
-e git+https://bitbucket.org/vinay.sajip/sarge#egg=sarge

#structlog
-e git+https://github.com/hynek/structlog#egg=structlog

#requests-cache
-e git+https://github.com/reclosedev/requests-cache#egg=requests-cache

#requests-html
-e git+https://github.com/psf/requests-html#egg=requests-html

#wrd_sphinx_theme
-e git+https://github.com/westurner/wrd_sphinx_theme#egg=wrd_sphinx_theme

matplotlib
